+++ tests/integration/test_attributes.py
@@ -25,19 +25,13 @@ def test_custom_attribute_report(mock_client_init):
 
     :param mock_client_init: Pytest fixture
     """
-    variables = {
-        'markers': 'scope: to which test scope a test relates'
-    }
+    variables = {'markers': 'scope: to which test scope a test relates'}
     variables.update(utils.DEFAULT_VARIABLES.items())
-    result = utils.run_pytest_tests(
-        tests=['examples/attributes/test_one_attribute.py'],
-        variables=variables
-    )
+    result = utils.run_pytest_tests(tests=['examples/attributes/test_one_attribute.py'], variables=variables)
     assert int(result) == 0, 'Exit code should be 0 (no errors)'
 
     mock_client = mock_client_init.return_value
-    assert mock_client.start_test_item.call_count > 0, \
-        '"start_test_item" called incorrect number of times'
+    assert mock_client.start_test_item.call_count > 0, '"start_test_item" called incorrect number of times'
 
     call_args = mock_client.start_test_item.call_args_list
     step_call_args = call_args[-1][1]
@@ -55,15 +49,11 @@ def test_custom_attribute_not_reported_if_skip_configured(mock_client_init):
         'rp_ignore_attributes': 'scope'
     }
     variables.update(utils.DEFAULT_VARIABLES.items())
-    result = utils.run_pytest_tests(
-        tests=['examples/attributes/test_one_attribute.py'],
-        variables=variables
-    )
+    result = utils.run_pytest_tests(tests=['examples/attributes/test_one_attribute.py'], variables=variables)
     assert int(result) == 0, 'Exit code should be 0 (no errors)'
 
     mock_client = mock_client_init.return_value
-    assert mock_client.start_test_item.call_count > 0, \
-        '"start_test_item" called incorrect number of times'
+    assert mock_client.start_test_item.call_count > 0, '"start_test_item" called incorrect number of times'
 
     call_args = mock_client.start_test_item.call_args_list
     step_call_args = call_args[-1][1]
@@ -76,19 +66,14 @@ def test_two_attributes_different_values_report(mock_client_init):
 
     :param mock_client_init: Pytest fixture
     """
-    variables = {
-        'markers': 'scope: to which test scope a test relates'
-    }
+    variables = {'markers': 'scope: to which test scope a test relates'}
     variables.update(utils.DEFAULT_VARIABLES.items())
     result = utils.run_pytest_tests(
-        tests=['examples/attributes/test_two_attributes_with_same_key.py'],
-        variables=variables
-    )
+        tests=['examples/attributes/test_two_attributes_with_same_key.py'], variables=variables)
     assert int(result) == 0, 'Exit code should be 0 (no errors)'
 
     mock_client = mock_client_init.return_value
-    assert mock_client.start_test_item.call_count > 0, \
-        '"start_test_item" called incorrect number of times'
+    assert mock_client.start_test_item.call_count > 0, '"start_test_item" called incorrect number of times'
 
     call_args = mock_client.start_test_item.call_args_list
     step_call_args = call_args[-1][1]
@@ -106,13 +91,11 @@ def test_skip_attribute(mock_client_init):
 
     :param mock_client_init: Pytest fixture
     """
-    result = utils.run_pytest_tests(
-        tests=['examples/skip/test_simple_skip.py'])
+    result = utils.run_pytest_tests(tests=['examples/skip/test_simple_skip.py'])
     assert int(result) == 0, 'Exit code should be 0 (no errors)'
 
     mock_client = mock_client_init.return_value
-    assert mock_client.start_test_item.call_count > 0, \
-        '"start_test_item" called incorrect number of times'
+    assert mock_client.start_test_item.call_count > 0, '"start_test_item" called incorrect number of times'
 
     call_args = mock_client.start_test_item.call_args_list
     step_call_args = call_args[-1][1]
@@ -129,22 +112,14 @@ def test_custom_runtime_attribute_report(mock_client_init):
 
     :param mock_client_init: Pytest fixture
     """
-    variables = {
-        'markers': 'scope: to which test scope a test relates\n'
-                   'runtime: runtime attribute mark'
-    }
+    variables = {'markers': 'scope: to which test scope a test relates\nruntime: runtime attribute mark'}
     variables.update(utils.DEFAULT_VARIABLES.items())
-    result = utils.run_pytest_tests(
-        tests=['examples/attributes/test_runtime_attribute.py'],
-        variables=variables
-    )
+    result = utils.run_pytest_tests(tests=['examples/attributes/test_runtime_attribute.py'], variables=variables)
     assert int(result) == 0, 'Exit code should be 0 (no errors)'
 
     mock_client = mock_client_init.return_value
-    assert mock_client.start_test_item.call_count > 0, \
-        '"start_test_item" called incorrect number of times'
-    assert mock_client.finish_test_item.call_count > 0, \
-        '"finish_test_item" called incorrect number of times'
+    assert mock_client.start_test_item.call_count > 0, '"start_test_item" called incorrect number of times'
+    assert mock_client.finish_test_item.call_count > 0, '"finish_test_item" called incorrect number of times'
 
     start_call_args = mock_client.start_test_item.call_args_list
     start_step_call_args = start_call_args[-1][1]